+++ changelog.md
@@ -3,17 +3,18 @@ This file contains all the notable changes done to the Ballerina Crypto package 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
 ## [Unreleased]
 
 ### Added
-
 - [Introduce Keccak-256 hashing algorithm](https://github.com/ballerina-platform/ballerina-library/issues/7509)
 
 ## [2.7.3] - 2025-01-20
 
 ### Added
-
 - [Introduce new APIs to support PGP encryption and decryption with streams](https://github.com/ballerina-platform/ballerina-library/issues/7064)
+- [Introduce new APIs to support Bcrypt and Argon2 hashing and verification](https://github.com/ballerina-platform/ballerina-library/issues/2744)
 
 ## [2.7.2] - 2024-05-30